requirements.txt
flet>=0.84.0
pypdf==4.2.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.1
